backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.3
django-admin==2.0.1
django-cors-headers==3.11.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.13.1
et-xmlfile==1.1.0
Flask==2.1.1
Flask-Cors==3.0.10
gcloud==0.18.3
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
oauth2client==4.1.3
openpyxl==3.0.9
pandas==1.4.1
Pillow==9.1.0
protobuf==3.20.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
screen==1.0.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
websockets==10.3
Werkzeug==2.1.1
XlsxWriter==3.0.3
xlwt==1.3.0
